+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Instance methods now accept native arrays of HTML elements.
 
 ### Changed
+- **Breaking:** The `reveal()` method no longer accepts an `interval` parameter. Instead, sequence intervals are now defined with `options.interval`.
 - **Breaking:** The `reveal()` method will not accept sequence intervals less than 16 milliseconds.
 - **Breaking:** The instance method `isSupported()` is now static.
 - **Breaking:** `options.distance` supports only `em` `px` and `%` values.